jordanjung-promptflow/requirements.txt
argcomplete
azure-ai-ml
azure-identity
azure-monitor-opentelemetry-exporter
azure-storage-blob
azureml-ai-monitoring
beautifulsoup4
bs4
colorama
cryptography
docutils
fastapi
filelock
filetype
flask
flask-cors
flask-restx
gitpython
httpx
jsonschema
keyring
marshmallow
openai
opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk
pandas
pillow
promptflow
psutil
pyarrow
pydash
pyinstaller
python-dateutil
python-dotenv
pywin32
requests
ruamel.yaml
sqlalchemy
streamlit
streamlit-quill
strictyaml
tabulate
tiktoken
waitress
PyPDF2
python-pptx
scikit-learn
